+++ src/NuvemFiscal.Sdk/Model/DocNFNFS.cs
@@ -39,7 +39,7 @@ namespace NuvemFiscal.Sdk.Model
         /// <param name="nNFS">Número da Nota Fiscal NF ou NFS. (required).</param>
         /// <param name="modNFS">Modelo da Nota Fiscal NF ou NFS. (required).</param>
         /// <param name="serieNFS">Série Nota Fiscal NF ou NFS. (required).</param>
-        public DocNFNFS(int? nNFS = default(int?), int? modNFS = default(int?), string serieNFS = default(string))
+        public DocNFNFS(int? nNFS = default(int?), long? modNFS = default(long?), string serieNFS = default(string))
         {
             // to ensure "nNFS" is required (not null)
             if (nNFS == null)
@@ -73,7 +73,7 @@ namespace NuvemFiscal.Sdk.Model
         /// </summary>
         /// <value>Modelo da Nota Fiscal NF ou NFS.</value>
         [DataMember(Name = "modNFS", IsRequired = true, EmitDefaultValue = true)]
-        public int? modNFS { get; set; }
+        public long? modNFS { get; set; }
 
         /// <summary>
         /// Série Nota Fiscal NF ou NFS.
